feat: split composable memory SDK workspaces
