+++ packages/contracts/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "@dsh-mnemon/contracts",
+  "version": "0.2.13",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  }
+}
