+++ packages/layer-documents/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@dsh-mnemon/layer-documents",
+  "version": "0.2.13",
+  "private": true,
+  "type": "module",
+  "exports": { ".": "./src/index.ts" }
+}
