+++ packages/layer-memory-spaces/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@dsh-mnemon/layer-memory-spaces",
+  "version": "0.2.13",
+  "private": true,
+  "type": "module",
+  "exports": { ".": "./src/index.ts" }
+}
