+++ packages/layer-runtime/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@dsh-mnemon/layer-runtime",
+  "version": "0.2.13",
+  "private": true,
+  "type": "module",
+  "exports": { ".": "./src/index.ts" }
+}
