+++ packages/provider-sdk/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@dsh-mnemon/provider-sdk",
+  "version": "0.2.13",
+  "private": true,
+  "type": "module",
+  "exports": { ".": "./src/index.ts" }
+}
