+++ packages/strategy-default-three-tier/package.json
@@ -1,0 +1,7 @@
+{
+  "name": "@dsh-mnemon/strategy-default-three-tier",
+  "version": "0.2.13",
+  "private": true,
+  "type": "module",
+  "exports": { ".": "./src/index.ts" }
+}
