+++ scripts/verify-package-contents.mjs
@@ -16,7 +16,23 @@ if (result.status !== 0) {
 
 const [pack] = JSON.parse(result.stdout)
 const paths = pack.files.map(file => file.path)
-const required = ['package.json', 'cordis.patch.yml', 'lib/index.js', 'lib/client.js', 'lib/types/index.d.ts', 'lib/types/client/index.d.ts']
+const required = [
+  'package.json',
+  'cordis.patch.yml',
+  'lib/index.js',
+  'lib/client.js',
+  'lib/contracts.js',
+  'lib/kernel.js',
+  'lib/extension-sdk.js',
+  'lib/provider-sdk.js',
+  'lib/strategy-sdk.js',
+  'lib/strategy-default-three-tier.js',
+  'lib/layers/runtime.js',
+  'lib/layers/documents.js',
+  'lib/layers/memory-spaces.js',
+  'lib/types/src/index.d.ts',
+  'lib/types/src/client/index.d.ts',
+]
 const allowedRootFiles = new Set(['package.json', 'cordis.patch.yml', 'LICENSE', 'README.md', 'README.zh-CN.md', 'SECURITY.md', 'THIRD_PARTY_NOTICES.md'])
 const missing = required.filter(path => !paths.includes(path))
 const unexpected = paths.filter(path => !allowedRootFiles.has(path) && !(/^lib\/.+\.(?:js|d\.ts)$/.test(path)))
@@ -34,12 +50,14 @@ const relativeReadmeImages = readmeFiles.flatMap((path) => {
     .map(source => `${path}: ${source}`)
 })
 
-if (missing.length > 0 || unexpected.length > 0 || hostLeaks.length > 0 || relativeReadmeImages.length > 0 || pack.unpackedSize > 1_500_000) {
+const maximumUnpackedBytes = 1_650_000
+
+if (missing.length > 0 || unexpected.length > 0 || hostLeaks.length > 0 || relativeReadmeImages.length > 0 || pack.unpackedSize > maximumUnpackedBytes) {
   if (missing.length > 0) console.error(`Missing package files:\n${missing.map(path => `- ${path}`).join('\n')}`)
   if (unexpected.length > 0) console.error(`Unexpected package files:\n${unexpected.map(path => `- ${path}`).join('\n')}`)
   if (hostLeaks.length > 0) console.error(`Host-only code leaked into lib/client.js:\n${hostLeaks.map(pattern => `- ${pattern}`).join('\n')}`)
   if (relativeReadmeImages.length > 0) console.error(`README images must use absolute HTTPS URLs so npm can render assets excluded from the package:\n${relativeReadmeImages.map(image => `- ${image}`).join('\n')}`)
-  if (pack.unpackedSize > 1_500_000) console.error(`Unpacked package is ${pack.unpackedSize} bytes; expected at most 1500000.`)
+  if (pack.unpackedSize > maximumUnpackedBytes) console.error(`Unpacked package is ${pack.unpackedSize} bytes; expected at most ${maximumUnpackedBytes}.`)
   process.exit(1)
 }
