+++ src/memory-system/access.ts
@@ -1,45 +1,2 @@
-import type {
-  MemoryCapability,
-  MemoryOperationTrigger,
-  MemoryParticipationChannel,
-  MemoryParticipationMode,
-  MemoryTopologyLayer,
-} from './contracts.ts'
-import { participationChannel } from './defaults.ts'
-
-export interface MemoryParticipationDecision {
-  allowed: boolean
-  layerId: string
-  capability: MemoryCapability
-  trigger: MemoryOperationTrigger
-  channel: MemoryParticipationChannel
-  mode: MemoryParticipationMode
-  reason?: string
-}
-
-/** Evaluate one topology boundary without invoking strategy or data-plane code. */
-export function decideMemoryLayerParticipation(
-  layer: MemoryTopologyLayer,
-  capability: MemoryCapability,
-  trigger: MemoryOperationTrigger,
-): MemoryParticipationDecision {
-  const channel = participationChannel(capability)
-  const mode = layer.participation[channel]
-  const base = { layerId: layer.id, capability, trigger, channel, mode }
-  if (!layer.enabled) return { ...base, allowed: false, reason: `memory layer ${layer.id} is disabled by the active topology` }
-  if (mode === 'off') return { ...base, allowed: false, reason: `memory layer ${layer.id} has ${channel} participation turned off` }
-  if (trigger !== 'manual' && mode !== 'automatic') {
-    return { ...base, allowed: false, reason: `memory layer ${layer.id} allows ${channel} only for manual operations` }
-  }
-  return { ...base, allowed: true }
-}
-
-export function assertMemoryLayerParticipation(
-  layer: MemoryTopologyLayer,
-  capability: MemoryCapability,
-  trigger: MemoryOperationTrigger,
-): MemoryParticipationDecision {
-  const decision = decideMemoryLayerParticipation(layer, capability, trigger)
-  if (!decision.allowed) throw new Error(decision.reason)
-  return decision
-}
+/** @deprecated Internal compatibility facade. Import from `dsh-mnemon/kernel`. */
+export * from '../../packages/kernel/src/access.ts'
