+++ src/provider-sdk.ts
@@ -1,0 +1,15 @@
+export * from '../packages/provider-sdk/src/index.ts'
+export {
+  MemoryProviderAdapterRegistry,
+  createBuiltinMemoryProviderAdapterRegistry,
+  memoryProviderAdapterFactories,
+  registerMemoryProviderAdapterFactory,
+} from './providers/registry.ts'
+export type { MemoryProviderAdapterFactory, MemoryProviderAdapterFactoryContext } from './providers/registry.ts'
+export type {
+  MemoryProviderAdapter,
+  ProviderBodyStatus,
+  ProviderMemorySpace,
+  ProviderScoreSemantics,
+  ProviderSearchResult,
+} from './providers/provider.ts'
