+++ tsconfig.types.json
@@ -6,10 +6,10 @@
     "declarationMap": false,
     "emitDeclarationOnly": true,
     "outDir": "lib/types",
-    "rootDir": "src",
+    "rootDir": ".",
     "allowImportingTsExtensions": false,
     "rewriteRelativeImportExtensions": true,
     "types": ["node"]
   },
-  "include": ["src/**/*.ts", "src/**/*.tsx"]
+  "include": ["src/**/*.ts", "src/**/*.tsx", "packages/**/*.ts"]
 }
